Refactor - Refatoração de funções javascripts  🔨
Refatoração de funções javascripts que estavam dificuldade a visualização do codigo
